+++ wandb/sdk/artifacts/storage_handlers/azure_handler.py
@@ -136,10 +136,11 @@ class AzureHandler(StorageHandler):
             name_starts_with=f"{blob_name}/"
         ):
             if len(entries) >= max_objects:
-                raise ValueError(
-                    f"Exceeded {max_objects} objects tracked, pass max_objects to "
-                    f"add_reference"
+                wandb.termwarn(
+                    f"Found more than {max_objects} objects under path, limiting upload "
+                    f"to {max_objects} objects. Increase max_objects to upload more"
                 )
+                break
             if not self._is_directory_stub(blob_properties):
                 suffix = PurePosixPath(blob_properties.name).relative_to(blob_name)
                 entries.append(
@@ -151,6 +152,7 @@ class AzureHandler(StorageHandler):
                         ),
                     )
                 )
+
         return entries
 
     def _get_module(self, name: str) -> ModuleType: